+++ SocialNetwork/src/persistence/PostDaoJDBC.java
@@ -31,7 +31,7 @@ public class PostDaoJDBC implements PostDao {
 			post.setDataCreazione(Calendar.getInstance().getTime());
 
 			String insert = "insert into post(id_post, email_utente, contenuto, "
-					+ "canale, gruppo, data_creazione) values (?,?,?,?,?,?)";
+					+ "canale, gruppo, data_creazione, numlikes,numdislikes) values (?,?,?,?,?,?,?,?)";
 
 			PreparedStatement statement = connection.prepareStatement(insert);
 			statement.setLong(1, post.getId());
@@ -40,6 +40,8 @@ public class PostDaoJDBC implements PostDao {
 			statement.setString(4, post.getCanale().getNome());
 			statement.setString(5, post.getGruppo().getNome());
 			statement.setTimestamp(6, new Timestamp(post.getDataCreazione().getTime()));
+			statement.setInt(7, post.getNumLikes());
+			statement.setInt(8, post.getNumDislikes());
 			statement.executeUpdate();
 			
 			this.updateLike(post, connection);
@@ -111,7 +113,8 @@ public class PostDaoJDBC implements PostDao {
 				post.setCanale(new CanaleDaoJDBC(dataSource).findByPrimaryKey(result.getString("canale")));
 				post.setGruppo(new GruppoDaoJDBC(dataSource).findByPrimaryKey(result.getString("gruppo"), post.getCanale().getNome()));
 				post.setDataCreazione(new Date(result.getTimestamp("data_creazione").getTime()));
-				
+				post.setNumLikes(result.getInt("numlikes"));
+				post.setNumDislikes(result.getInt("numDislikes"));
 			}
 		} catch (SQLException e) {
 			throw new PersistenceException(e.getMessage());
@@ -142,7 +145,9 @@ public class PostDaoJDBC implements PostDao {
 				post.setCanale(new CanaleDaoJDBC(dataSource).findByPrimaryKey(result.getString("canale")));
 				post.setGruppo(new GruppoDaoJDBC(dataSource).findByPrimaryKey(result.getString("gruppo"), post.getCanale().getNome()));
 				post.setDataCreazione(new Date(result.getTimestamp("data_creazione").getTime()));
-
+				post.setNumLikes(result.getInt("numlikes"));
+				post.setNumDislikes(result.getInt("numDislikes"));
+			
 				allPost.add(post);
 			}
 		} catch (SQLException e) {
@@ -264,12 +269,23 @@ public class PostDaoJDBC implements PostDao {
 			statement.setString(2, utente.getEmail());
 			ResultSet result = statement.executeQuery();
 			
-			if(!result.next()) {
-				
+			if(!result.next()) {				
 				String insert = "insert into likes(id_post, email_utente) values (?,?)";
 				statement = connection.prepareStatement(insert);
 				statement.setLong(1, post.getId());
 				statement.setString(2, utente.getEmail());
+				statement.executeUpdate();
+				
+				statement = connection.prepareStatement("Select numlikes from post where id_post = ?");
+				statement.setLong(1, post.getId());
+				result = statement.executeQuery();
+				
+				result.next();
+				int n = result.getInt(1)+1;
+				
+				statement = connection.prepareStatement("update post SET numlikes = ? WHERE id_post = ?");
+				statement.setLong(2, post.getId());
+				statement.setInt(1, n);
 				statement.executeUpdate();
 			}
 
@@ -294,12 +310,30 @@ public class PostDaoJDBC implements PostDao {
 		
 		Connection connection = dataSource.getConnection();
 		try {
-			String delete = "delete FROM likes WHERE id_post = ? and email_utente = ?";
-			PreparedStatement statement = connection.prepareStatement(delete);
+			PreparedStatement statement = connection.prepareStatement("Select * from likes where id_post = ? and email_utente = ?");
 			statement.setLong(1, post.getId());
 			statement.setString(2, utente.getEmail());
+			ResultSet result = statement.executeQuery();
+			
+			if(result.next()) {			
+				String delete = "delete FROM likes WHERE id_post = ? and email_utente = ?";
+				statement = connection.prepareStatement(delete);
+				statement.setLong(1, post.getId());
+				statement.setString(2, utente.getEmail());
+				statement.executeUpdate();
+				
+				statement = connection.prepareStatement("Select numlikes from post where id_post = ?");
+				statement.setLong(1, post.getId());
+				result = statement.executeQuery();
 
-			statement.executeUpdate();
+				result.next();
+				int n = result.getInt(1)-1;
+				
+				statement = connection.prepareStatement("update post SET numlikes = ? WHERE id_post = ?");
+				statement.setLong(2, post.getId());
+				statement.setInt(1, n);
+				statement.executeUpdate();
+			}
 		} catch (SQLException e) {
 			if (connection != null)
 				try {
@@ -326,12 +360,23 @@ public class PostDaoJDBC implements PostDao {
 			statement.setString(2, utente.getEmail());
 			ResultSet result = statement.executeQuery();
 			
-			if(!result.next()) {
-				
+			if(!result.next()) {				
 				String insert = "insert into dislikes(id_post, email_utente) values (?,?)";
 				statement = connection.prepareStatement(insert);
 				statement.setLong(1, post.getId());
 				statement.setString(2, utente.getEmail());
+				statement.executeUpdate();
+				
+				statement = connection.prepareStatement("Select numdislikes from post where id_post = ?");
+				statement.setLong(1, post.getId());
+				result = statement.executeQuery();
+
+				result.next();
+				int n = result.getInt(1)+1;
+				
+				statement = connection.prepareStatement("update post SET numdislikes = ? WHERE id_post = ?");
+				statement.setLong(2, post.getId());
+				statement.setInt(1, n);
 				statement.executeUpdate();
 			}
 
@@ -356,12 +401,30 @@ public class PostDaoJDBC implements PostDao {
 		
 		Connection connection = dataSource.getConnection();
 		try {
-			String delete = "delete FROM dislikes WHERE id_post = ? and email_utente = ?";
-			PreparedStatement statement = connection.prepareStatement(delete);
+			PreparedStatement statement = connection.prepareStatement("Select * from dislikes where id_post = ? and email_utente = ?");
 			statement.setLong(1, post.getId());
 			statement.setString(2, utente.getEmail());
-
-			statement.executeUpdate();
+			ResultSet result = statement.executeQuery();
+			
+			if(result.next()) {
+				String delete = "delete FROM dislikes WHERE id_post = ? and email_utente = ?";
+				statement = connection.prepareStatement(delete);
+				statement.setLong(1, post.getId());
+				statement.setString(2, utente.getEmail());
+				statement.executeUpdate();
+				
+				statement = connection.prepareStatement("Select numdislikes from post where id_post = ?");
+				statement.setLong(1, post.getId());
+				result = statement.executeQuery();
+	
+				result.next();
+				int n = result.getInt(1)-1;
+				
+				statement = connection.prepareStatement("update post SET numdislikes = ? WHERE id_post = ?");
+				statement.setLong(2, post.getId());
+				statement.setInt(1, n);
+				statement.executeUpdate();
+			}
 		} catch (SQLException e) {
 			if (connection != null)
 				try {
